datacapture/ingestion/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
bitarray==3.0.0
cchecksum==0.0.4
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
clickhouse-connect==0.8.13
comm==0.2.2
cytoolz==1.0.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
dune_client==1.7.7
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
eth_retry==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
frozenlist==1.5.0
grpclib==0.4.7
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lz4==4.3.3
markdown-it-py==3.0.0
marshmallow==3.25.0
matplotlib-inline==0.1.7
mdurl==0.1.2
modal==0.72.2
multicall==0.10.0
multidict==6.1.0
mypy-extensions==1.0.0
ndjson==0.3.1
nest-asyncio==1.6.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
polars==1.19.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.0.1
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
synchronicity==0.9.8
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.1
types-certifi==2021.10.8.3
types-Deprecated==1.2.15.20241117
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
types-setuptools==75.6.0.20241223
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchfiles==1.0.3
wcwidth==0.2.13
web3==7.6.1
websockets==13.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
